% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_mean.R
\name{idx_mean}
\alias{idx_mean}
\title{Create Mean Index from Several Variables}
\usage{
idx_mean(..., fill_na = TRUE, na.rm = TRUE)
}
\arguments{
\item{...}{One or more unquoted variables}

\item{fill_na}{Option to fill missing values from any variable with linear predictions from all of the other variables. \code{TRUE} by default.}

\item{na.rm}{Option to remove NA's when calculating mean index. \code{TRUE} by default.}
}
\value{
A vector with the mean index
}
\description{
Create Mean Index from Several Variables
}
\examples{

var1 <- rnorm(100, mean = 0.5, sd = 0.25)
var2 <- rnorm(100, mean = -1, sd = 0.5)
var3 <- rnorm(100, mean = 1, sd = 1)

idx_mean(var1, var2, var3)

library(dplyr)

df <- tibble(var1, var2, var3)

mutate(df, idx_var = idx_mean(var1, var2, var3))

}
