% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_idx.R
\name{mean_idx}
\alias{mean_idx}
\title{Create Mean Index from Several Variables}
\usage{
mean_idx(...)
}
\arguments{
\item{...}{One or more unquoted variables}
}
\value{

}
\description{
Create Mean Index from Several Variables
}
