% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize.R
\name{stdize}
\alias{stdize}
\title{Standardize Variable}
\usage{
stdize(variable, compare, condition, na.rm = TRUE)
}
\arguments{
\item{variable}{Unquoted name of variable to standardize}

\item{compare}{Variable to use to calculate mean and standard deviation from}

\item{condition}{Logical vector (TRUE or FALSE) used to subset \code{compare} variable before calculating mean and standard deviation.}

\item{na.rm}{Option to remove NA's when calculating mean and standard deviation. Set to \code{TRUE} by default. Note that all values will be NA in the standardized variable if set to \code{FALSE}.}
}
\value{
A vector with the standardized variable
}
\description{
Standardize Variable
}
\examples{

var <- rnorm(100, mean = 0.5, sd = 1)

var_std <- stdize(var)

summary(var_std)

}
