% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_idx.R
\name{mean_idx}
\alias{mean_idx}
\title{Create Mean Index from Several Variables}
\usage{
mean_idx(...)
}
\arguments{
\item{...}{One or more unquoted variables}
}
\value{

}
\description{
Create Mean Index from Several Variables
}
\examples{

var1 <- rnorm(100, mean = 0.5, sd = 0.25)
var2 <- rnorm(100, mean = -1, sd = 0.5)
var3 <- rnorm(100, mean = 1, sd = 1)

mean_idx(var1, var2, var3)

}
