% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize.R
\name{stdize}
\alias{stdize}
\title{Standardize Variable}
\usage{
stdize(variable, compare, condition, na.rm = TRUE)
}
\arguments{
\item{variable}{Unquoted name of variable to standardize}

\item{compare}{Variable to use to calculate mean and standard deviation from}

\item{condition}{Logical vector (TRUE or FALSE) used to subset \code{compare} variable before calculating mean and standard deviation.}

\item{na.rm}{}
}
\value{

}
\description{
Standardize Variable
}
